+++ src/main/kotlin/com/eitanliu/dart/mappable/actions/JsonToDartAction.kt
@@ -1,8 +1,11 @@
 package com.eitanliu.dart.mappable.actions
 
-import com.eitanliu.dart.mappable.extensions.invokeLater
+import com.eitanliu.intellij.compat.extensions.ApplicationScope
+import com.eitanliu.intellij.compat.extensions.value
 import com.eitanliu.dart.mappable.ui.JsonInputDialog
+import com.eitanliu.dart.mappable.utils.ApplicationUtils
 import com.eitanliu.dart.mappable.utils.CommandUtils
+import com.eitanliu.dart.mappable.utils.DependenciesUtils
 import com.eitanliu.dart.mappable.utils.MessagesUtils
 import com.intellij.CommonBundle
 import com.intellij.codeInsight.actions.ReformatCodeProcessor
@@ -10,7 +13,6 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.LangDataKeys
 import com.intellij.openapi.actionSystem.PlatformDataKeys
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.editor.Document
 import com.intellij.openapi.fileEditor.FileDocumentManager
@@ -22,6 +24,9 @@ import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
 import com.intellij.psi.impl.file.PsiDirectoryFactory
 import io.flutter.pub.PubRoot
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.future.await
+import kotlinx.coroutines.launch
 import java.io.IOException
 
 
@@ -95,8 +100,20 @@ class JsonToDartAction : AnAction() {
             }
         }
 
-        ApplicationManager.getApplication().invokeLater(600L) {
+        // ApplicationManager.getApplication().invokeLater(600L) {
+        ApplicationScope.launch {
+            delay(600L)
+
+            val moduleEntity = DependenciesUtils.loadModule(project, module, pubRoot)
+            val settings = ApplicationUtils.getSettings()
+            val deps = DependenciesUtils.checkDependencies(settings, moduleEntity.pubspec)
+            // println("deps ${deps.joinToString()}")
+            for (dep in deps) {
+                CommandUtils.executeFlutterPubCommand(project, pubRoot, dep).await()
+            }
+
             FileDocumentManager.getInstance().saveDocument(doc)
+            if (!settings.graph.autoBuildRunner.value) return@launch
             CommandUtils.executeFlutterPubCommand(
                 project, pubRoot, "run build_runner build --delete-conflicting-outputs"
             ) {