+++ src/main/kotlin/com/eitanliu/dart/mappable/ast/DartFileName.kt
@@ -19,10 +19,10 @@ interface DartFileName {
 
     val fileName: String
 
-    class Default(
+    class Impl(
         className: String,
         suffix: String,
-    ) : DartFileName, Extensions {
+    ) : Self {
         override val camelCaseName = className.keyToCamelCase(true)
 
         override val camelCaseSuffix = suffix
@@ -40,11 +40,18 @@ interface DartFileName {
         override val fileName = "$underscoreNameAndSuffix.dart"
     }
 
-    interface Extensions {
+    // self members and extension functions
+    interface Self : DartFileName, Ext
 
-        fun String.keyToCamelCase(capitalizeFirstWord: Boolean = false): String {
+    interface Ext {
+
+        fun String.keyToFieldName(firstWord: Boolean = false): String {
+            return trimStart('_').keyToCamelCase(firstWord)
+        }
+
+        fun String.keyToCamelCase(firstWord: Boolean = false): String {
             return replaceNonAlphabeticNumber("_")
-                .underscoreToCamelCase(capitalizeFirstWord)
+                .underscoreToCamelCase(firstWord)
                 .numberAddPrefix()
         }