+++ src/main/kotlin/com/eitanliu/dart/mappable/generator/DartGeneratorBuilder.kt
@@ -1,7 +1,7 @@
 package com.eitanliu.dart.mappable.generator
 
 import com.eitanliu.dart.mappable.ast.DartGenerator
-import com.eitanliu.dart.mappable.extensions.value
+import com.eitanliu.intellij.compat.extensions.value
 import com.eitanliu.dart.mappable.settings.Implements
 import com.eitanliu.dart.mappable.settings.Settings
 
@@ -11,5 +11,7 @@ fun buildDartGenerator(
     json: String,
 ): DartGenerator = when (settings.graph.implement.value) {
     Implements.JSON_SERIALIZABLE -> JsonSerializableGenerator(settings, className, json)
-    else -> DartMappableGenerator(settings, className, json)
+    Implements.DART_MAPPABLE -> DartMappableGenerator(settings, className, json)
+    Implements.FREEZED -> FreezedGenerator(settings, className, json)
+    else -> JsonSerializableGenerator(settings, className, json)
 }