+++ src/main/kotlin/com/eitanliu/dart/mappable/settings/SettingLayout.kt
@@ -2,9 +2,9 @@ package com.eitanliu.dart.mappable.settings
 
 import com.eitanliu.dart.mappable.binding.bindSelected
 import com.eitanliu.dart.mappable.binding.selected
-import com.eitanliu.dart.mappable.extensions.createPropertyGraph
-import com.eitanliu.dart.mappable.extensions.propertyOf
-import com.eitanliu.dart.mappable.extensions.value
+import com.eitanliu.intellij.compat.extensions.createPropertyGraph
+import com.eitanliu.intellij.compat.extensions.propertyOf
+import com.eitanliu.intellij.compat.extensions.value
 import com.intellij.openapi.observable.properties.PropertyGraph
 import com.intellij.openapi.options.UnnamedConfigurable
 import com.intellij.ui.TitledSeparator
@@ -32,7 +32,29 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
 
             rowComment("Configure dart data model files suffix.")
         }
+        row {
+            checkBox(
+                "auto run build_runner"
+            ).bindSelected(
+                graph.autoBuildRunner
+            ).applyToComponent {
+                toolTipText = "auto run 'flutter pub run build_runner build --delete-conflicting-outputs'"
+            }
+        }
         buttonsGroup("Implement") {
+            row {
+                checkBox(
+                    "json_reflectable"
+                ).bindSelected(
+                    graph.enableJsonReflectable
+                ).applyToComponent {
+                    toolTipText = "use json_reflectable annotation"
+                }
+            }
+            row {
+                radioButton("json_serializable", Implements.JSON_SERIALIZABLE)
+                    .bindSelected(graph.implement, Implements.JSON_SERIALIZABLE)
+            }
             val mappablePredicate = graph.implement.selected(Implements.DART_MAPPABLE)
             row {
                 radioButton("dart_mappable", Implements.DART_MAPPABLE)
@@ -42,10 +64,14 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
             rowsRange {
                 buildMappable()
             }.visibleIf(mappablePredicate)
+            val freezedPredicate = graph.implement.selected(Implements.FREEZED)
             row {
-                radioButton("json_serializable", Implements.JSON_SERIALIZABLE)
-                    .bindSelected(graph.implement, Implements.JSON_SERIALIZABLE)
+                radioButton("freezed", Implements.FREEZED)
+                    .bindSelected(graph.implement, Implements.FREEZED)
             }
+            rowsRange {
+                buildFreezed()
+            }.visibleIf(freezedPredicate)
         }.bind(::implement)
 
     }
@@ -90,15 +116,24 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
                         bindText(graph.mappableToJson)
                     }.align(AlignX.FILL)
                 }.layout(RowLayout.LABEL_ALIGNED)
-                // row {
-                //     checkBox("copyWith")
-                //         .bindSelected(graph.enableCopyWith)
-                //     textField().apply {
-                //         bindText(graph.mappableCopyWith)
-                //     }.align(AlignX.FILL)
-                // }.layout(RowLayout.LABEL_ALIGNED)
+                row {
+                    checkBox("copyWith")
+                        .bindSelected(graph.enableCopyWith)
+                    textField().apply {
+                        bindText(graph.mappableCopyWith)
+                    }.align(AlignX.FILL)
+                }.layout(RowLayout.LABEL_ALIGNED)
             }.visibleIf(customPredicate)
         }.bind(::enableMixin)
+    }
+
+    private fun Panel.buildFreezed() {
+        indent {
+            row {
+                checkBox("fromJson/toJson")
+                    .bindSelected(graph.freezedEnableJson)
+            }
+        }
     }
 
     private fun Panel.rowPanel(
@@ -121,6 +156,8 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
     override fun isModified(): Boolean {
         return settings.graph.modelSuffix.value != graph.modelSuffix.value
                 || settings.graph.implement.value != graph.implement.value
+                || settings.graph.autoBuildRunner.value != graph.autoBuildRunner.value
+                || settings.graph.enableJsonReflectable.value != graph.enableJsonReflectable.value
                 || settings.graph.enableMixin.value != graph.enableMixin.value
                 || settings.graph.enableFromJson.value != graph.enableFromJson.value
                 || settings.graph.enableToJson.value != graph.enableToJson.value
@@ -132,11 +169,14 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
                 || settings.graph.mappableFromMap.value != graph.mappableFromMap.value
                 || settings.graph.mappableToMap.value != graph.mappableToMap.value
                 || settings.graph.mappableCopyWith.value != graph.mappableCopyWith.value
+                || settings.graph.freezedEnableJson.value != graph.freezedEnableJson.value
     }
 
     override fun apply() {
         settings.graph.modelSuffix.value = graph.modelSuffix.value
         settings.graph.implement.value = graph.implement.value
+        settings.graph.autoBuildRunner.value = graph.autoBuildRunner.value
+        settings.graph.enableJsonReflectable.value = graph.enableJsonReflectable.value
         settings.graph.enableMixin.value = graph.enableMixin.value
         settings.graph.enableFromJson.value = graph.enableFromJson.value
         settings.graph.enableToJson.value = graph.enableToJson.value
@@ -148,11 +188,14 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
         settings.graph.mappableFromMap.value = graph.mappableFromMap.value
         settings.graph.mappableToMap.value = graph.mappableToMap.value
         settings.graph.mappableCopyWith.value = graph.mappableCopyWith.value
+        settings.graph.freezedEnableJson.value = graph.freezedEnableJson.value
     }
 
     override fun reset() {
         graph.modelSuffix.value = settings.graph.modelSuffix.value
         graph.implement.value = settings.graph.implement.value
+        graph.autoBuildRunner.value = settings.graph.autoBuildRunner.value
+        graph.enableJsonReflectable.value = settings.graph.enableJsonReflectable.value
         graph.enableMixin.value = settings.graph.enableMixin.value
         graph.enableFromJson.value = settings.graph.enableFromJson.value
         graph.enableToJson.value = settings.graph.enableToJson.value
@@ -164,6 +207,7 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
         graph.mappableFromMap.value = settings.graph.mappableFromMap.value
         graph.mappableToMap.value = settings.graph.mappableToMap.value
         graph.mappableCopyWith.value = settings.graph.mappableCopyWith.value
+        graph.freezedEnableJson.value = settings.graph.freezedEnableJson.value
     }
 
     class Graph(data: SettingLayout) {
@@ -171,6 +215,8 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
 
         val modelSuffix = propertyGraph.propertyOf(data.settings.modelSuffix)
         val implement = propertyGraph.propertyOf(data.settings.implement)
+        val autoBuildRunner = propertyGraph.propertyOf(data.settings.autoBuildRunner)
+        val enableJsonReflectable = propertyGraph.propertyOf(data.settings.enableJsonReflectable)
         val enableMixin = propertyGraph.propertyOf(data.settings.enableMixin)
         val enableFromJson = propertyGraph.propertyOf(data.settings.enableFromJson)
         val enableToJson = propertyGraph.propertyOf(data.settings.enableToJson)
@@ -182,6 +228,7 @@ class SettingLayout(private val settings: Settings) : UnnamedConfigurable {
         val mappableFromMap = propertyGraph.propertyOf(data.settings.mappableFromMap)
         val mappableToMap = propertyGraph.propertyOf(data.settings.mappableToMap)
         val mappableCopyWith = propertyGraph.propertyOf(data.settings.mappableCopyWith)
+        val freezedEnableJson = propertyGraph.propertyOf(data.settings.freezedEnableJson)
 
     }
 }