+++ src/main/kotlin/com/eitanliu/intellij/compat/application/CoroutineSupport.kt
@@ -1,0 +1,11 @@
+package com.eitanliu.intellij.compat.application
+
+import com.intellij.openapi.application.ModalityState
+import kotlin.coroutines.CoroutineContext
+
+interface CoroutineSupport {
+
+    fun asContextElement(modalityState: ModalityState): CoroutineContext
+
+    fun edtDispatcher(): CoroutineContext
+}