+++ src/main/kotlin/com/eitanliu/intellij/compat/extensions/FunctionExt.kt
@@ -1,0 +1,8 @@
+package com.eitanliu.intellij.compat.extensions
+
+import com.eitanliu.intellij.compat.ref.WeakFun1Reference
+
+fun <T> (T.() -> Unit).weakRef(): T.() -> Unit {
+    return WeakFun1Reference(this)
+}
+