+++ src/main/kotlin/com/eitanliu/intellij/compat/extensions/PropertyGraphExt.kt
@@ -1,13 +1,13 @@
 @file:Suppress("unused", "NOTHING_TO_INLINE")
 
-package com.eitanliu.dart.mappable.extensions
+package com.eitanliu.intellij.compat.extensions
 
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.observable.properties.GraphProperty
 import com.intellij.openapi.observable.properties.PropertyGraph
+import java.lang.reflect.Constructor
 import kotlin.reflect.KMutableProperty0
 import kotlin.reflect.KProperty0
-import kotlin.reflect.full.createInstance
 
 inline fun <V> KProperty0<V>.toGraphProperty(
     propertyGraph: PropertyGraph = createPropertyGraph()
@@ -25,14 +25,34 @@ inline fun <V> KMutableProperty0<V>.toGraphProperty(
 fun createPropertyGraph() = try {
     PropertyGraph()
 } catch (e: Throwable) {
-    PropertyGraph::class.createInstance()
+    PropertyGraph::class.java.createInstance({ sequence ->
+        sequence.sortedByDescending { it.parameterCount }
+            .first { it.parameterCount <= 2 }
+    }) {
+        when (it.parameterCount) {
+            2 -> arrayOf(null, true)
+            else -> null
+        }
+    }
 }
+
+@Suppress("UNCHECKED_CAST")
+fun <T> Class<T>.createInstance(
+    predicate: ((sequence: Sequence<Constructor<*>>) -> Constructor<*>?)? = null,
+    params: ((Constructor<*>) -> Array<*>?)? = null,
+): T = run {
+    val sequence = constructors.asSequence()
+    val constructor = predicate?.invoke(sequence) ?: sequence.first()
+    val args = params?.invoke(constructor) ?: arrayOfNulls<Any?>(constructors[0].parameterCount)
+    constructor.newInstance(*args) as T
+}
+
 
 inline fun <T> PropertyGraph.propertyRef(ref: KProperty0<T>): GraphProperty<T> = ref.toGraphProperty(this)
 
 inline fun <T> PropertyGraph.propertyRef(ref: KMutableProperty0<T>): GraphProperty<T> = ref.toGraphProperty(this)
 
-fun <T> PropertyGraph.propertyOf(initial: T): GraphProperty<T> = property(initial)
+inline fun <T> PropertyGraph.propertyOf(initial: T): GraphProperty<T> = propertyOf { initial }
 
 fun <T> PropertyGraph.propertyOf(initial: () -> T): GraphProperty<T> = lazyProperty(initial)