+++ src/main/kotlin/com/eitanliu/intellij/compat/ref/WeakFunReference.kt
@@ -1,0 +1,15 @@
+package com.eitanliu.intellij.compat.ref
+
+import java.lang.ref.WeakReference
+
+class WeakFun1Reference<T>(
+    func: T.() -> Unit
+) : Function1<T, Unit> {
+
+    private val func = WeakReference(func)
+
+    override fun invoke(p1: T) {
+        func.get()?.invoke(p1)
+    }
+
+}